requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
Django==4.2
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework==3.14.0
executing==1.2.0
idna==3.4
ipdb==0.13.13
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytz==2023.3
redis==4.5.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
urllib3==1.26.15
uWSGI==2.0.21
wcwidth==0.2.6
git+https://github.com/jeagercoder/service_lib.git@v0.0.4
